pyproject.toml
[project]
name = "evalml"
version = "0.1.0-beta.1"
description = "Evaluation CLI for ML models"
readme = "README.md"
authors = [
    { name = "Francesco Zanetta", email = "francesco.zanetta@meteoswiss.ch" },
    { name = "Daniele Nerini", email = "daniele.nerini@meteoswiss.ch" },
]
requires-python = ">=3.11"
dependencies = [
    "snakemake<9.10", # https://github.com/snakemake/snakemake/issues/3289
    "snakemake-executor-plugin-slurm",
    "click",
    "meteodata-lab>=0.4.0",
    "anemoi-datasets>=0.5.25",
    "mlflow>=3.1.1",
    "pydantic>=2.11.7",
    "toml>=0.10.2",
    "netcdf4>=1.7.2",
    "cartopy",
    "earthkit-plots",
    "marimo>=0.16.5",
]

[project.optional-dependencies]
kerchunk = [
    "kerchunk",
    "zarr<3.0.0",
    "fastparquet",
    "ujson"
]

[project.scripts]
evalml = "evalml:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "snakefmt>=0.11.0",
]

[tool.pytest.ini_options]
markers = [
    "longtest: mark tests that take a long time to run, e.g. integration tests",
]

[tool.hatch.build.targets.wheel]
packages = [
  "src/evalml",
  "src/verification"
]
